remove sequence.h
