+++ cpp/src/arrow/python/python_to_arrow.cc
@@ -28,7 +28,7 @@
 #include "arrow/python/numpy_convert.h"
 #include "arrow/python/numpy_interop.h"
 #include "arrow/python/platform.h"
-#include "arrow/python/sequence.h"
+#include "arrow/python/python_to_arrow-internal.h"
 
 constexpr int32_t kMaxRecursionDepth = 100;
 
@@ -39,61 +39,6 @@ PyObject* pyarrow_deserialize_callback = NULL;
 
 namespace arrow {
 namespace py {
-
-/// Constructing dictionaries of key/value pairs. Sequences of
-/// keys and values are built separately using a pair of
-/// SequenceBuilders. The resulting Arrow representation
-/// can be obtained via the Finish method.
-class DictBuilder {
- public:
-  explicit DictBuilder(MemoryPool* pool = nullptr) : keys_(pool), vals_(pool) {}
-
-  /// Builder for the keys of the dictionary
-  SequenceBuilder& keys() { return keys_; }
-  /// Builder for the values of the dictionary
-  SequenceBuilder& vals() { return vals_; }
-
-  /// Construct an Arrow StructArray representing the dictionary.
-  /// Contains a field "keys" for the keys and "vals" for the values.
-
-  /// \param list_data
-  ///    List containing the data from nested lists in the value
-  ///   list of the dictionary
-  ///
-  /// \param dict_data
-  ///   List containing the data from nested dictionaries in the
-  ///   value list of the dictionary
-  arrow::Status Finish(std::shared_ptr<Array> key_tuple_data,
-                       std::shared_ptr<Array> key_dict_data,
-                       std::shared_ptr<Array> val_list_data,
-                       std::shared_ptr<Array> val_tuple_data,
-                       std::shared_ptr<Array> val_dict_data, std::shared_ptr<Array>* out);
-
- private:
-  SequenceBuilder keys_;
-  SequenceBuilder vals_;
-};
-
-Status DictBuilder::Finish(std::shared_ptr<Array> key_tuple_data,
-                           std::shared_ptr<Array> key_dict_data,
-                           std::shared_ptr<Array> val_list_data,
-                           std::shared_ptr<Array> val_tuple_data,
-                           std::shared_ptr<Array> val_dict_data,
-                           std::shared_ptr<Array>* out) {
-  // lists and dicts can't be keys of dicts in Python, that is why for
-  // the keys we do not need to collect sublists
-  std::shared_ptr<Array> keys, vals;
-  RETURN_NOT_OK(keys_.Finish(nullptr, key_tuple_data, key_dict_data, &keys));
-  RETURN_NOT_OK(vals_.Finish(val_list_data, val_tuple_data, val_dict_data, &vals));
-  auto keys_field = std::make_shared<Field>("keys", keys->type());
-  auto vals_field = std::make_shared<Field>("vals", vals->type());
-  auto type = std::make_shared<StructType>(
-      std::vector<std::shared_ptr<Field>>({keys_field, vals_field}));
-  std::vector<std::shared_ptr<Array>> field_arrays({keys, vals});
-  DCHECK(keys->length() == vals->length());
-  out->reset(new StructArray(type, keys->length(), field_arrays));
-  return Status::OK();
-}
 
 Status CallCustomCallback(PyObject* callback, PyObject* elem, PyObject** result) {
   *result = NULL;
