+++ cpp/src/arrow/python/python_to_arrow.h
@@ -24,7 +24,6 @@
 #include "arrow/io/interfaces.h"
 #include "arrow/python/numpy_interop.h"
 #include "arrow/python/platform.h"
-#include "arrow/python/sequence.h"
 
 #include <vector>
